+++ histoire.config.ts
@@ -2,6 +2,7 @@ import { defineConfig } from 'histoire'
 import { HstVue } from '@histoire/plugin-vue'
 
 export default defineConfig({
+  setupFile: '/src/historie.setup.ts',
   plugins: [
     HstVue(),
   ],